Fix a float issue, rendering works again - but a problem with enemy states

+++ src/audio_visual/gui/hud.c
@@ -49,8 +49,8 @@ inlinable void draw_minimap(const vec pos) {
 	SDL_FRect wall_tile = {0.0, origin[1], scale[0], scale[1]};
 	Color color = hud_colors[HUD];
 
-	for (int map_y = 0; map_y < tile_amts.y; map_y++, wall_tile.y += scale[1], wall_tile.x = 0.0) {
-		for (int map_x = 0; map_x < tile_amts.x; map_x++, wall_tile.x += scale[0]) {
+	for (int map_y = 0; map_y < tile_amts.y; map_y++, wall_tile.y += (float) scale[1], wall_tile.x = 0.0) {
+		for (int map_x = 0; map_x < tile_amts.x; map_x++, wall_tile.x += (float) scale[0]) {
 
 			const byte point = map_point(current_level.wall_data, map_x, map_y);
 			const byte point_height = current_level.get_point_height(point, (vec) {map_x, map_y});
